Module Tjr_profile

Profiling is controlled via optcomp flags. The Optcomp_config module records those values as OCaml values.

Example (see Example module in Tjr_profile):

let f () = 
  (* do something *)
  1+2

(** waypoint *)
let pt = intern "pt"

let profiler = make_profiler ()
let {mark;_} = profiler

let profiled_f () = 
  mark pt;
  let r = f () in
  mark (-1*pt);
  r

(** Run f 100 times, and print the profiling results at exit *)
let do_it () = 
  for i = 0 to 100 do
    ignore(profiled_f())
  done

Configuation

include module type of struct include Tjr_profile_core.Optcomp_config end

This is an optcomp configuration file. It imports a raw optcomp file from "config.ml" which defines two optcomp variables: PROFILING_USE_TSC and PROFILING_ENABLED. The rest of the file is optcomp-filtered OCaml.

The underlying timing method is controlled by optcomp PROFILING_USE_TSC variable.

val profiling_use_tsc : bool

We ARE using TSC for profiling.

val now : unit -> int

val profiling_enabled : bool

Profiling IS enabled (for profilers)

Interface

module Profile_intf : sig ... end

Profiling interfaces

include module type of struct include Profile_intf end

type 'waypoint profiler = {

mark : 'waypoint -> unit;

get_marks : unit -> ('waypoint * int) list;

print_summary : unit -> unit;

time_thunk : a. 'waypoint -> (unit -> 'a) -> 'a;

}

Profiler ops:

mark: imperative mark command

(pvt) get_marks: a list of all marked waypoints and the corresponding times at which they were marked

print_summary: print a human-readable summary (see Pretty_print)

val dummy_profiler : 'a profiler

exception Profiling_exception of string

module With_array : sig ... end

This version attempts to reduce computation and GC costs during the profiling (there may be substantial processing costs at the end to compute the summary).

val make_profiler : ?print_header:string -> ?cap:int -> ?print_at_exit:bool -> unit -> int Tjr_profile__Intf_.profiler

By default, we use the array-based profiler

val intern : string -> int

Waypoints are int (for efficiency); strings can be interned to convert them to ints; when interning a string "foo", we also intern the string "foo'" as the negative of the intern of foo; this allows a common "entry/exit" idiom

Profile a single function

module Profile_single_function : sig ... end

Profile a single function which may be called many times, using "enter" and "exit". This code keeps track of the number of times, and the total time.

val measure_execution_time_and_print : string -> (unit -> 'a) -> 'a

Measure execution time, and print result on output immediately. This is not affected by optcomp config.

Example

module Example () : sig ... end