Module Tjr_profile.Profile_single_function

Profile a single function which may be called many times, using "enter" and "exit". This code keeps track of the number of times, and the total time.

type stats = {

total : int;

count : int;

}

type profile_single = {

enter : unit -> unit;

exit : unit -> unit;

get_stats : unit -> stats;

}

val make : ?print_at_exit:bool -> ?print_header:string -> unit -> profile_single