Module Tjr_profile__.Util

module Make_intern : functor (S : sig ... end) -> sig ... end

Intern (allocate an integer for) objects FIXME move to tjr_lib

module Make_intern_2 : functor (S : sig ... end) -> sig ... end

module Intern : sig ... end

val measure_execution_time_and_print : string -> (unit -> 'a) -> 'a

Measure execution time, and print result on output immediately. This is not affected by optcomp config.

val measure_execution_time : (unit -> 'a) -> 'a Tjr_profile__.Intf_.Timed_result.timed_result

Measure the execution time of a function, and return the function result and the execution time. This is not affected by optcomp config.

val dummy_profiler : 'a Tjr_profile__.Intf_.profiler