Module Private.Small

Small implementation

module Make_1 : sig ... end

module Make_2 : Small_intf.T

Functionality to create a named timer, get the current time, and add a given duration to a timer

module Make_3 : Small_intf.T

Trivial stub implementation, which doesn't do anything;