Module Tjr_profile

Profiling is controlled via optcomp flags. The Optcomp_config module records those values as OCaml values.

include Tjr_profile__.Optcomp_config

val profiling_use_tsc : bool

We ARE using TSC for profiling.

val now : unit -> Core_kernel__.Import.int

val profiling_enabled : bool

Profiling IS enabled (for profilers)

module Profile_intf = Tjr_profile__.Intf_

include Profile_intf

type 'waypoint profiler = {

mark : 'waypoint -> unit;

time_thunk : a. 'waypoint -> (unit -> 'a) -> 'a;

get_marks : unit -> ('waypoint * int) list;

print_summary : unit -> unit;

}

Profiler ops:

mark: imperative mark command

get_marks: a list of all marked waypoints and the corresponding times at which they were marked

print_summary: print a human-readable summary (see Pretty_print)

type stats = {

total : int;

count : int;

}

type profile_single = {

enter : unit -> unit;

exit : unit -> unit;

get_stats : unit -> stats;

}

module Timed_result = Tjr_profile__.Intf_.Timed_result

type 'a timed_result = 'a Timed_result.timed_result

exception Profiling_exception of string

include Tjr_profile__.Util

module Make_intern = Tjr_profile__.Util.Make_intern

Intern (allocate an integer for) objects FIXME move to tjr_lib

module Make_intern_2 = Tjr_profile__.Util.Make_intern_2

module Intern = Tjr_profile__.Util.Intern

val measure_execution_time_and_print : string -> (unit -> 'a) -> 'a

Measure execution time, and print result on output immediately. This is not affected by optcomp config.

val measure_execution_time : (unit -> 'a) -> 'a Tjr_profile__.Intf_.Timed_result.timed_result

Measure the execution time of a function, and return the function result and the execution time. This is not affected by optcomp config.

val dummy_profiler : 'a Tjr_profile__.Intf_.profiler

module Profile_single_function : sig ... end

Profile a single function using "enter" and "exit". The function may be called many times. This keeps track of the number of times, and the total time.

module With_array : sig ... end

val intern : string -> int

val make_profiler : ?⁠print_header:string -> ?⁠cap:int -> ?⁠print_at_exit:bool -> unit -> int Tjr_profile__.Intf_.profiler

val dummy_profiler : 'a Tjr_profile__.Intf_.profiler